Finish auth
