+++ web/router.ex
@@ -30,36 +30,15 @@ defmodule WhatIf.Router do
   def verify_jwt(conn, opts) do
     case Plug.Conn.get_req_header(conn, "authorization") do
       ["Bearer " <> token] ->
-        kid = get_kid(token)
-        firebase_keys = get_public_keys()
-        case Map.get(firebase_keys, kid) do
-          jwk = %JOSE.JWK{} ->
-            verified = JOSE.JWT.verify_strict(jwk, ["RS256"], token)
-            case verified do
-              {true, fields, _} ->
-                assign(conn, :user_id, fields.fields["user_id"])
-              _ ->
-                conn |> unauthorized
-            end
-          e ->
+        case WhatIf.FirebaseVeryfier.verify_jwt(token) do
+          {true, user_id} ->
+            assign(conn, :user_id, user_id)
+          _ ->
             conn |> unauthorized
         end
       _ ->
         conn |> unauthorized
     end
-  end
-
-  defp get_kid(token) do
-    try do
-      token
-      |> JOSE.JWS.peek_protected()
-      |> JOSE.decode()
-      |> Map.get("kid")
-    catch
-      e, f ->
-        nil
-    end
-
   end
 
   defp unauthorized(conn) do
@@ -70,12 +49,7 @@ defmodule WhatIf.Router do
   end
 
 
-  defp get_public_keys() do
-        {:ok, {{_, 200, _}, _, body}} = :httpc.request(:get, {'https://www.googleapis.com/robot/v1/metadata/x509/securetoken@system.gserviceaccount.com', []}, [autoredirect: true], [])
-    JOSE.JWK.from_firebase(IO.iodata_to_binary(body))
-  end
-
-  # Other scopes may use custom stacks.
+    # Other scopes may use custom stacks.
   # scope "/api", WhatIf do
   #   pipe_through :api
   # end
